python/src/wallets/multiversx/pyproject.toml
[tool.poetry]
name = "goat-sdk-wallet-multiversx"
version = "0.1.0"
description = "Goat SDK MultiversX wallet implementation"
authors = [
    "Ofem Eteng <ofemeteng@gmail.com>",
]
readme = "README.md"
keywords = [
    "goat",
    "sdk",
    "multiversx",
    "agents",
    "ai",
    "wallet",
]
homepage = "https://ohmygoat.dev/"
repository = "https://github.com/goat-sdk/goat"
packages = [
    { include = "goat_wallets/multiversx" },
]

[tool.poetry.dependencies]
python = "^3.10"
goat-sdk = "^0.1.0"
multiversx-sdk = ">=1.1.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.6"

[tool.poetry.group.dev.dependencies.goat-sdk]
path = "../../goat-sdk"
develop = true

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/goat-sdk/goat/issues"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = "src"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 120
target-version = "py312"

[build-system]
requires = [
    "poetry-core",
]
build-backend = "poetry.core.masonry.api"
